+++ metta/map/scenes/inline_ascii.py
@@ -31,20 +31,20 @@ class InlineAscii(Scene):
     def __init__(self, data: str, row: int = 0, column: int = 0):
         super().__init__()
         lines = data.strip().splitlines()
-        self._level = np.array([list(line) for line in lines], dtype="U6")
-        self._level = np.vectorize(SYMBOLS.get)(self._level)
+        self._grid = np.array([list(line) for line in lines], dtype="U6")
+        self._grid = np.vectorize(SYMBOLS.get)(self._grid)
         self._row = row
         self._column = column
 
     def _render(self, node):
-        if node.width < self._level.shape[1] + self._column or node.height < self._level.shape[0] + self._row:
+        if node.width < self._grid.shape[1] + self._column or node.height < self._grid.shape[0] + self._row:
             raise ValueError(
-                f"Level size {self._level.shape} is too large for node size {node.width}x{node.height} at "
+                f"Grid size {self._grid.shape} is too large for node size {node.width}x{node.height} at "
                 f"{self._column},{self._row}"
             )
 
-        level_height, level_width = self._level.shape
+        grid_height, grid_width = self._grid.shape
         node.grid[
-            self._row : self._row + level_height,
-            self._column : self._column + level_width,
-        ] = self._level
+            self._row : self._row + grid_height,
+            self._column : self._column + grid_width,
+        ] = self._grid
